Added control functions to simulation files, cleaned and commented simulation code.
Cleaned and commented Particle files.

Added viewController files so I can move glut functionality into its own object instead of main

+++ Physics_Integrator/Physics_Integrator/Integrator.h
@@ -18,7 +18,7 @@ using namespace std;
 class Integrator {
 public:
 	Integrator();
-	void integrate_verlet(glm::vec3 *pos, glm::vec3 *last_pos, float ts, float m, glm::vec3 force);
+	void integrate_verlet(glm::vec3 *pos, glm::vec3 *last_pos, double ts, float m, glm::vec3 force);
 	void integrate_euler();
 };
